+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 12.0.0-beta.12 (Unreleased)
 
+### Breaking Changes
+
+- Renamed `ContentLength` in `FlushFileResult` to `FileSize`.
 
 ## 12.0.0-beta.11 (2021-05-19)